Struct aws_sdk_eks::client::fluent_builders::CreateNodegroup

pub struct CreateNodegroup<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateNodegroup<C>

pub async fn send(
self
) -> Result<CreateNodegroupOutput, SdkError<CreateNodegroupError>> where
C: SmithyConnector,

pub fn cluster_name(self, input: impl Into<String>) -> Self

The name of the cluster to create the node group in.

pub fn set_cluster_name(self, input: Option<String>) -> Self

pub fn nodegroup_name(self, input: impl Into<String>) -> Self

The unique name to give your node group.

pub fn set_nodegroup_name(self, input: Option<String>) -> Self

pub fn scaling_config(self, input: NodegroupScalingConfig) -> Self

The scaling configuration details for the Auto Scaling group that is created for your node group.

pub fn set_scaling_config(self, input: Option<NodegroupScalingConfig>) -> Self

pub fn disk_size(self, input: i32) -> Self

The root device disk size (in GiB) for your node group instances. The default disk size is 20 GiB. If you specify launchTemplate, then don't specify diskSize, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_disk_size(self, input: Option<i32>) -> Self

pub fn subnets(self, inp: impl Into<String>) -> Self

The subnets to use for the Auto Scaling group that is created for your node group. If you specify launchTemplate, then don't specify SubnetId in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_subnets(self, input: Option<Vec<String>>) -> Self

pub fn instance_types(self, inp: impl Into<String>) -> Self

Specify the instance types for a node group. If you specify a GPU instance type, be sure to specify AL2_x86_64_GPU with the amiType parameter. If you specify launchTemplate, then you can specify zero or one instance type in your launch template or you can specify 0-20 instance types for instanceTypes. If however, you specify an instance type in your launch template and specify any instanceTypes, the node group deployment will fail. If you don't specify an instance type in a launch template or for instanceTypes, then t3.medium is used, by default. If you specify Spot for capacityType, then we recommend specifying multiple values for instanceTypes. For more information, see Managed node group capacity types and Launch template support in the Amazon EKS User Guide.

pub fn set_instance_types(self, input: Option<Vec<String>>) -> Self

pub fn ami_type(self, input: AmiTypes) -> Self

The AMI type for your node group. GPU instance types should use the AL2_x86_64_GPU AMI type. Non-GPU instances should use the AL2_x86_64 AMI type. Arm instances should use the AL2_ARM_64 AMI type. All types use the Amazon EKS optimized Amazon Linux 2 AMI. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify amiType, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_ami_type(self, input: Option<AmiTypes>) -> Self

pub fn remote_access(self, input: RemoteAccessConfig) -> Self

The remote access (SSH) configuration to use with your node group. If you specify launchTemplate, then don't specify remoteAccess, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_remote_access(self, input: Option<RemoteAccessConfig>) -> Self

pub fn node_role(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role to associate with your node group. The Amazon EKS worker node kubelet daemon makes calls to AWS APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies. Before you can launch nodes and register them into a cluster, you must create an IAM role for those nodes to use when they are launched. For more information, see Amazon EKS node IAM role in the Amazon EKS User Guide . If you specify launchTemplate, then don't specify IamInstanceProfile in your launch template, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_node_role(self, input: Option<String>) -> Self

pub fn labels(self, k: impl Into<String>, v: impl Into<String>) -> Self

The Kubernetes labels to be applied to the nodes in the node group when they are created.

pub fn set_labels(self, input: Option<HashMap<String, String>>) -> Self

pub fn taints(self, inp: impl Into<Taint>) -> Self

The Kubernetes taints to be applied to the nodes in the node group.

pub fn set_taints(self, input: Option<Vec<Taint>>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

The metadata to apply to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn launch_template(self, input: LaunchTemplateSpecification) -> Self

An object representing a node group's launch template specification. If specified, then do not specify instanceTypes, diskSize, or remoteAccess and make sure that the launch template meets the requirements in launchTemplateSpecification.

pub fn set_launch_template(
self,
input: Option<LaunchTemplateSpecification>
) -> Self

pub fn update_config(self, input: NodegroupUpdateConfig) -> Self

pub fn set_update_config(self, input: Option<NodegroupUpdateConfig>) -> Self

pub fn capacity_type(self, input: CapacityTypes) -> Self

The capacity type for your node group.

pub fn set_capacity_type(self, input: Option<CapacityTypes>) -> Self

pub fn version(self, input: impl Into<String>) -> Self

The Kubernetes version to use for your managed nodes. By default, the Kubernetes version of the cluster is used, and this is the only accepted specified value. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify version, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_version(self, input: Option<String>) -> Self

pub fn release_version(self, input: impl Into<String>) -> Self

The AMI version of the Amazon EKS optimized AMI to use with your node group. By default, the latest available AMI version for the node group's current Kubernetes version is used. For more information, see Amazon EKS optimized Amazon Linux 2 AMI versions in the Amazon EKS User Guide. If you specify launchTemplate, and your launch template uses a custom AMI, then don't specify releaseVersion, or the node group deployment will fail. For more information about using launch templates with Amazon EKS, see Launch template support in the Amazon EKS User Guide.

pub fn set_release_version(self, input: Option<String>) -> Self