Struct aws_sdk_eks::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSWesleyFrontend.

This client allows ergonomic access to a AWSWesleyFrontend-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn associate_encryption_config(&self) -> AssociateEncryptionConfig<C, M, R>

Constructs a fluent builder for the AssociateEncryptionConfig operation.

See AssociateEncryptionConfig for more information about the operation and its arguments.

pub fn associate_identity_provider_config(
&self
) -> AssociateIdentityProviderConfig<C, M, R>

Constructs a fluent builder for the AssociateIdentityProviderConfig operation.

See AssociateIdentityProviderConfig for more information about the operation and its arguments.

pub fn create_addon(&self) -> CreateAddon<C, M, R>

Constructs a fluent builder for the CreateAddon operation.

See CreateAddon for more information about the operation and its arguments.

pub fn create_cluster(&self) -> CreateCluster<C, M, R>

Constructs a fluent builder for the CreateCluster operation.

See CreateCluster for more information about the operation and its arguments.

pub fn create_fargate_profile(&self) -> CreateFargateProfile<C, M, R>

Constructs a fluent builder for the CreateFargateProfile operation.

See CreateFargateProfile for more information about the operation and its arguments.

pub fn create_nodegroup(&self) -> CreateNodegroup<C, M, R>

Constructs a fluent builder for the CreateNodegroup operation.

See CreateNodegroup for more information about the operation and its arguments.

pub fn delete_addon(&self) -> DeleteAddon<C, M, R>

Constructs a fluent builder for the DeleteAddon operation.

See DeleteAddon for more information about the operation and its arguments.

pub fn delete_cluster(&self) -> DeleteCluster<C, M, R>

Constructs a fluent builder for the DeleteCluster operation.

See DeleteCluster for more information about the operation and its arguments.

pub fn delete_fargate_profile(&self) -> DeleteFargateProfile<C, M, R>

Constructs a fluent builder for the DeleteFargateProfile operation.

See DeleteFargateProfile for more information about the operation and its arguments.

pub fn delete_nodegroup(&self) -> DeleteNodegroup<C, M, R>

Constructs a fluent builder for the DeleteNodegroup operation.

See DeleteNodegroup for more information about the operation and its arguments.

pub fn deregister_cluster(&self) -> DeregisterCluster<C, M, R>

Constructs a fluent builder for the DeregisterCluster operation.

See DeregisterCluster for more information about the operation and its arguments.

pub fn describe_addon(&self) -> DescribeAddon<C, M, R>

Constructs a fluent builder for the DescribeAddon operation.

See DescribeAddon for more information about the operation and its arguments.

pub fn describe_addon_versions(&self) -> DescribeAddonVersions<C, M, R>

Constructs a fluent builder for the DescribeAddonVersions operation.

See DescribeAddonVersions for more information about the operation and its arguments.

pub fn describe_cluster(&self) -> DescribeCluster<C, M, R>

Constructs a fluent builder for the DescribeCluster operation.

See DescribeCluster for more information about the operation and its arguments.

pub fn describe_fargate_profile(&self) -> DescribeFargateProfile<C, M, R>

Constructs a fluent builder for the DescribeFargateProfile operation.

See DescribeFargateProfile for more information about the operation and its arguments.

pub fn describe_identity_provider_config(
&self
) -> DescribeIdentityProviderConfig<C, M, R>

Constructs a fluent builder for the DescribeIdentityProviderConfig operation.

See DescribeIdentityProviderConfig for more information about the operation and its arguments.

pub fn describe_nodegroup(&self) -> DescribeNodegroup<C, M, R>

Constructs a fluent builder for the DescribeNodegroup operation.

See DescribeNodegroup for more information about the operation and its arguments.

pub fn describe_update(&self) -> DescribeUpdate<C, M, R>

Constructs a fluent builder for the DescribeUpdate operation.

See DescribeUpdate for more information about the operation and its arguments.

pub fn disassociate_identity_provider_config(
&self
) -> DisassociateIdentityProviderConfig<C, M, R>

Constructs a fluent builder for the DisassociateIdentityProviderConfig operation.

See DisassociateIdentityProviderConfig for more information about the operation and its arguments.

pub fn list_addons(&self) -> ListAddons<C, M, R>

Constructs a fluent builder for the ListAddons operation.

See ListAddons for more information about the operation and its arguments.

pub fn list_clusters(&self) -> ListClusters<C, M, R>

Constructs a fluent builder for the ListClusters operation.

See ListClusters for more information about the operation and its arguments.

pub fn list_fargate_profiles(&self) -> ListFargateProfiles<C, M, R>

Constructs a fluent builder for the ListFargateProfiles operation.

See ListFargateProfiles for more information about the operation and its arguments.

pub fn list_identity_provider_configs(
&self
) -> ListIdentityProviderConfigs<C, M, R>

Constructs a fluent builder for the ListIdentityProviderConfigs operation.

See ListIdentityProviderConfigs for more information about the operation and its arguments.

pub fn list_nodegroups(&self) -> ListNodegroups<C, M, R>

Constructs a fluent builder for the ListNodegroups operation.

See ListNodegroups for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_updates(&self) -> ListUpdates<C, M, R>

Constructs a fluent builder for the ListUpdates operation.

See ListUpdates for more information about the operation and its arguments.

pub fn register_cluster(&self) -> RegisterCluster<C, M, R>

Constructs a fluent builder for the RegisterCluster operation.

See RegisterCluster for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_addon(&self) -> UpdateAddon<C, M, R>

Constructs a fluent builder for the UpdateAddon operation.

See UpdateAddon for more information about the operation and its arguments.

pub fn update_cluster_config(&self) -> UpdateClusterConfig<C, M, R>

Constructs a fluent builder for the UpdateClusterConfig operation.

See UpdateClusterConfig for more information about the operation and its arguments.

pub fn update_cluster_version(&self) -> UpdateClusterVersion<C, M, R>

Constructs a fluent builder for the UpdateClusterVersion operation.

See UpdateClusterVersion for more information about the operation and its arguments.

pub fn update_nodegroup_config(&self) -> UpdateNodegroupConfig<C, M, R>

Constructs a fluent builder for the UpdateNodegroupConfig operation.

See UpdateNodegroupConfig for more information about the operation and its arguments.

pub fn update_nodegroup_version(&self) -> UpdateNodegroupVersion<C, M, R>

Constructs a fluent builder for the UpdateNodegroupVersion operation.

See UpdateNodegroupVersion for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.