Module aws_sdk_eks::error

Modules

bad_request_exception

See BadRequestException

client_exception

See ClientException

invalid_parameter_exception

See InvalidParameterException

invalid_request_exception

See InvalidRequestException

not_found_exception

See NotFoundException

resource_in_use_exception

See ResourceInUseException

resource_limit_exceeded_exception

See ResourceLimitExceededException

resource_not_found_exception

See ResourceNotFoundException

server_exception

See ServerException

service_unavailable_exception

See ServiceUnavailableException

unsupported_availability_zone_exception

See UnsupportedAvailabilityZoneException

Structs

AssociateEncryptionConfigError

AssociateIdentityProviderConfigError

BadRequestException

This exception is thrown if the request contains a semantic error. The precise meaning

ClientException

These errors are usually caused by a client action. Actions can include using an

CreateAddonError

CreateClusterError

CreateFargateProfileError

CreateNodegroupError

DeleteAddonError

DeleteClusterError

DeleteFargateProfileError

DeleteNodegroupError

DescribeAddonError

DescribeAddonVersionsError

DescribeClusterError

DescribeFargateProfileError

DescribeIdentityProviderConfigError

DescribeNodegroupError

DescribeUpdateError

DisassociateIdentityProviderConfigError

InvalidParameterException

The specified parameter is invalid. Review the available parameters for the API

InvalidRequestException

The request is invalid given the state of the cluster. Check the state of the cluster

ListAddonsError

ListClustersError

ListFargateProfilesError

ListIdentityProviderConfigsError

ListNodegroupsError

ListTagsForResourceError

ListUpdatesError

NotFoundException

A service resource associated with the request could not be found. Clients should not

ResourceInUseException

The specified resource is in use.

ResourceLimitExceededException

You have encountered a service limit on the specified resource.

ResourceNotFoundException

The specified resource could not be found. You can view your available clusters with

ServerException

These errors are usually caused by a server-side issue.

ServiceUnavailableException

The service is unavailable. Back off and retry the operation.

TagResourceError

UnsupportedAvailabilityZoneException

At least one of your specified cluster subnets is in an Availability Zone that does

UntagResourceError

UpdateAddonError

UpdateClusterConfigError

UpdateClusterVersionError

UpdateNodegroupConfigError

UpdateNodegroupVersionError

Enums

AssociateEncryptionConfigErrorKind

AssociateIdentityProviderConfigErrorKind

CreateAddonErrorKind

CreateClusterErrorKind

CreateFargateProfileErrorKind

CreateNodegroupErrorKind

DeleteAddonErrorKind

DeleteClusterErrorKind

DeleteFargateProfileErrorKind

DeleteNodegroupErrorKind

DescribeAddonErrorKind

DescribeAddonVersionsErrorKind

DescribeClusterErrorKind

DescribeFargateProfileErrorKind

DescribeIdentityProviderConfigErrorKind

DescribeNodegroupErrorKind

DescribeUpdateErrorKind

DisassociateIdentityProviderConfigErrorKind

ListAddonsErrorKind

ListClustersErrorKind

ListFargateProfilesErrorKind

ListIdentityProviderConfigsErrorKind

ListNodegroupsErrorKind

ListTagsForResourceErrorKind

ListUpdatesErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAddonErrorKind

UpdateClusterConfigErrorKind

UpdateClusterVersionErrorKind

UpdateNodegroupConfigErrorKind

UpdateNodegroupVersionErrorKind