Module aws_sdk_eks::model

Modules

addon

See Addon

addon_health

See AddonHealth

addon_info

See AddonInfo

addon_issue

See AddonIssue

addon_version_info

See AddonVersionInfo

auto_scaling_group

See AutoScalingGroup

certificate

See Certificate

cluster

See Cluster

compatibility

See Compatibility

encryption_config

See EncryptionConfig

error_detail

See ErrorDetail

fargate_profile

See FargateProfile

fargate_profile_selector

See FargateProfileSelector

identity

See Identity

identity_provider_config

See IdentityProviderConfig

identity_provider_config_response

See IdentityProviderConfigResponse

issue

See Issue

kubernetes_network_config_request

See KubernetesNetworkConfigRequest

kubernetes_network_config_response

See KubernetesNetworkConfigResponse

launch_template_specification

See LaunchTemplateSpecification

log_setup

See LogSetup

logging

See Logging

nodegroup

See Nodegroup

nodegroup_health

See NodegroupHealth

nodegroup_resources

See NodegroupResources

nodegroup_scaling_config

See NodegroupScalingConfig

nodegroup_update_config

See NodegroupUpdateConfig

oidc

See Oidc

oidc_identity_provider_config

See OidcIdentityProviderConfig

oidc_identity_provider_config_request

See OidcIdentityProviderConfigRequest

provider

See Provider

remote_access_config

See RemoteAccessConfig

taint

See Taint

update

See Update

update_labels_payload

See UpdateLabelsPayload

update_param

See UpdateParam

update_taints_payload

See UpdateTaintsPayload

vpc_config_request

See VpcConfigRequest

vpc_config_response

See VpcConfigResponse

Structs

Addon

An Amazon EKS add-on.

AddonHealth

The health of the add-on.

AddonInfo

Information about an add-on.

AddonIssue

An issue related to an add-on.

AddonVersionInfo

Information about an add-on version.

AutoScalingGroup

An Auto Scaling group that is associated with an Amazon EKS managed node group.

Certificate

An object representing the certificate-authority-data for your

Cluster

An object representing an Amazon EKS cluster.

Compatibility

Compatibility information.

EncryptionConfig

The encryption configuration for the cluster.

ErrorDetail

An object representing an error when an asynchronous operation fails.

FargateProfile

An object representing an AWS Fargate profile.

FargateProfileSelector

An object representing an AWS Fargate profile selector.

Identity

An object representing an identity provider.

IdentityProviderConfig

An object representing an identity provider configuration.

IdentityProviderConfigResponse

An object that represents an identity configuration.

Issue

An object representing an issue with an Amazon EKS resource.

KubernetesNetworkConfigRequest

The Kubernetes network configuration for the cluster.

KubernetesNetworkConfigResponse

The Kubernetes network configuration for the cluster.

LaunchTemplateSpecification

An object representing a node group launch template specification. The launch template

LogSetup

An object representing the enabled or disabled Kubernetes control plane logs for your

Logging

An object representing the logging configuration for resources in your cluster.

Nodegroup

An object representing an Amazon EKS managed node group.

NodegroupHealth

An object representing the health status of the node group.

NodegroupResources

An object representing the resources associated with the node group, such as Auto

NodegroupScalingConfig

An object representing the scaling configuration details for the Auto Scaling group

NodegroupUpdateConfig

Oidc

An object representing the OpenID

OidcIdentityProviderConfig

An object that represents the configuration for an OpenID Connect (OIDC) identity

OidcIdentityProviderConfigRequest

An object representing an OpenID Connect (OIDC) configuration. Before associating an

Provider

Identifies the AWS Key Management Service (AWS KMS) key used to encrypt the secrets.

RemoteAccessConfig

An object representing the remote access configuration for the managed node

Taint

A property that allows a node to repel a set of pods.

Update

An object representing an asynchronous update.

UpdateLabelsPayload

An object representing a Kubernetes label change for a managed node group.

UpdateParam

An object representing the details of an update request.

UpdateTaintsPayload

An object representing the details of an update to a taints payload.

VpcConfigRequest

An object representing the VPC configuration to use for an Amazon EKS cluster.

VpcConfigResponse

An object representing an Amazon EKS cluster VPC configuration response.

Enums

AddonIssueCode

AddonStatus

AmiTypes

CapacityTypes

ClusterStatus

ConfigStatus

ErrorCode

NOTE: ErrorCode::Unknown has been renamed to ::UnknownValue.

FargateProfileStatus

LogType

NodegroupIssueCode

NodegroupStatus

ResolveConflicts

TaintEffect

UpdateParamType

UpdateStatus

UpdateType