Struct aws_sdk_eks::model::cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Cluster

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the cluster.

pub fn set_name(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the cluster.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The Unix epoch timestamp in seconds for when the cluster was created.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn version(self, input: impl Into<String>) -> Self

The Kubernetes server version for the cluster.

pub fn set_version(self, input: Option<String>) -> Self

pub fn endpoint(self, input: impl Into<String>) -> Self

The endpoint for your Kubernetes API server.

pub fn set_endpoint(self, input: Option<String>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM role that provides permissions for the Kubernetes control plane to make calls to AWS API operations on your behalf.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn resources_vpc_config(self, input: VpcConfigResponse) -> Self

The VPC configuration used by the cluster control plane. Amazon EKS VPC resources have specific requirements to work properly with Kubernetes. For more information, see Cluster VPC Considerations and Cluster Security Group Considerations in the Amazon EKS User Guide.

pub fn set_resources_vpc_config(self, input: Option<VpcConfigResponse>) -> Self

pub fn kubernetes_network_config(
self,
input: KubernetesNetworkConfigResponse
) -> Self

The Kubernetes network configuration for the cluster.

pub fn set_kubernetes_network_config(
self,
input: Option<KubernetesNetworkConfigResponse>
) -> Self

pub fn logging(self, input: Logging) -> Self

The logging configuration for your cluster.

pub fn set_logging(self, input: Option<Logging>) -> Self

pub fn identity(self, input: Identity) -> Self

The identity provider information for the cluster.

pub fn set_identity(self, input: Option<Identity>) -> Self

pub fn status(self, input: ClusterStatus) -> Self

The current status of the cluster.

pub fn set_status(self, input: Option<ClusterStatus>) -> Self

pub fn certificate_authority(self, input: Certificate) -> Self

The certificate-authority-data for your cluster.

pub fn set_certificate_authority(self, input: Option<Certificate>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn platform_version(self, input: impl Into<String>) -> Self

The platform version of your Amazon EKS cluster. For more information, see Platform Versions in the Amazon EKS User Guide .

pub fn set_platform_version(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn encryption_config(self, input: impl Into<EncryptionConfig>) -> Self

pub fn set_encryption_config(self, input: Option<Vec<EncryptionConfig>>) -> Self

pub fn build(self) -> Cluster

Consumes the builder and constructs a Cluster