Struct aws_sdk_eks::model::oidc_identity_provider_config_request::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for OidcIdentityProviderConfigRequest

Implementations

impl Builder

pub fn identity_provider_config_name(self, input: impl Into<String>) -> Self

The name of the OIDC provider configuration.

pub fn set_identity_provider_config_name(self, input: Option<String>) -> Self

pub fn issuer_url(self, input: impl Into<String>) -> Self

The URL of the OpenID identity provider that allows the API server to discover public signing keys for verifying tokens. The URL must begin with https:// and should correspond to the iss claim in the provider's OIDC ID tokens. Per the OIDC standard, path components are allowed but query parameters are not. Typically the URL consists of only a hostname, like https://server.example.org or https://example.com. This URL should point to the level below .well-known/openid-configuration and must be publicly accessible over the internet.

pub fn set_issuer_url(self, input: Option<String>) -> Self

pub fn client_id(self, input: impl Into<String>) -> Self

This is also known as audience. The ID for the client application that makes authentication requests to the OpenID identity provider.

pub fn set_client_id(self, input: Option<String>) -> Self

pub fn username_claim(self, input: impl Into<String>) -> Self

The JSON Web Token (JWT) claim to use as the username. The default is sub, which is expected to be a unique identifier of the end user. You can choose other claims, such as email or name, depending on the OpenID identity provider. Claims other than email are prefixed with the issuer URL to prevent naming clashes with other plug-ins.

pub fn set_username_claim(self, input: Option<String>) -> Self

pub fn username_prefix(self, input: impl Into<String>) -> Self

The prefix that is prepended to username claims to prevent clashes with existing names. If you do not provide this field, and username is a value other than email, the prefix defaults to issuerurl#. You can use the value - to disable all prefixing.

pub fn set_username_prefix(self, input: Option<String>) -> Self

pub fn groups_claim(self, input: impl Into<String>) -> Self

The JWT claim that the provider uses to return your groups.

pub fn set_groups_claim(self, input: Option<String>) -> Self

pub fn groups_prefix(self, input: impl Into<String>) -> Self

The prefix that is prepended to group claims to prevent clashes with existing names (such as system: groups). For example, the value oidc: will create group names like oidc:engineering and oidc:infra.

pub fn set_groups_prefix(self, input: Option<String>) -> Self

pub fn required_claims(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_required_claims(self, input: Option<HashMap<String, String>>) -> Self

pub fn build(self) -> OidcIdentityProviderConfigRequest

Consumes the builder and constructs a OidcIdentityProviderConfigRequest