Struct aws_sdk_eks::model::Nodegroup

#[non_exhaustive]
pub struct Nodegroup {

    pub nodegroup_name: Option<String>,
    pub nodegroup_arn: Option<String>,
    pub cluster_name: Option<String>,
    pub version: Option<String>,
    pub release_version: Option<String>,
    pub created_at: Option<Instant>,
    pub modified_at: Option<Instant>,
    pub status: Option<NodegroupStatus>,
    pub capacity_type: Option<CapacityTypes>,
    pub scaling_config: Option<NodegroupScalingConfig>,
    pub instance_types: Option<Vec<String>>,
    pub subnets: Option<Vec<String>>,
    pub remote_access: Option<RemoteAccessConfig>,
    pub ami_type: Option<AmiTypes>,
    pub node_role: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub taints: Option<Vec<Taint>>,
    pub resources: Option<NodegroupResources>,
    pub disk_size: Option<i32>,
    pub health: Option<NodegroupHealth>,
    pub update_config: Option<NodegroupUpdateConfig>,
    pub launch_template: Option<LaunchTemplateSpecification>,
    pub tags: Option<HashMap<String, String>>,

}

An object representing an Amazon EKS managed node group.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

nodegroup_name: Option<String>

The name associated with an Amazon EKS managed node group.

nodegroup_arn: Option<String>

The Amazon Resource Name (ARN) associated with the managed node group.

cluster_name: Option<String>

The name of the cluster that the managed node group resides in.

version: Option<String>

The Kubernetes version of the managed node group.

release_version: Option<String>

If the node group was deployed using a launch template with a custom AMI, then this is the AMI ID that was specified in the launch template. For node groups that weren't deployed using a launch template, this is the version of the Amazon EKS optimized AMI that the node group was deployed with.

created_at: Option<Instant>

The Unix epoch timestamp in seconds for when the managed node group was created.

modified_at: Option<Instant>

The Unix epoch timestamp in seconds for when the managed node group was last modified.

status: Option<NodegroupStatus>

The current status of the managed node group.

capacity_type: Option<CapacityTypes>

The capacity type of your managed node group.

scaling_config: Option<NodegroupScalingConfig>

The scaling configuration details for the Auto Scaling group that is associated with your node group.

instance_types: Option<Vec<String>>

If the node group wasn't deployed with a launch template, then this is the instance type that is associated with the node group. If the node group was deployed with a launch template, then this is null.

subnets: Option<Vec<String>>

The subnets that were specified for the Auto Scaling group that is associated with your node group.

remote_access: Option<RemoteAccessConfig>

If the node group wasn't deployed with a launch template, then this is the remote access configuration that is associated with the node group. If the node group was deployed with a launch template, then this is null.

ami_type: Option<AmiTypes>

If the node group was deployed using a launch template with a custom AMI, then this is CUSTOM. For node groups that weren't deployed using a launch template, this is the AMI type that was specified in the node group configuration.

node_role: Option<String>

The IAM role associated with your node group. The Amazon EKS node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies.

labels: Option<HashMap<String, String>>

The Kubernetes labels applied to the nodes in the node group.

Only labels that are applied with the Amazon EKS API are shown here. There may be other Kubernetes labels applied to the nodes in this group.

taints: Option<Vec<Taint>>

The Kubernetes taints to be applied to the nodes in the node group when they are created. Effect is one of No_Schedule, Prefer_No_Schedule, or No_Execute. Kubernetes taints can be used together with tolerations to control how workloads are scheduled to your nodes.

resources: Option<NodegroupResources>

The resources associated with the node group, such as Auto Scaling groups and security groups for remote access.

disk_size: Option<i32>

If the node group wasn't deployed with a launch template, then this is the disk size in the node group configuration. If the node group was deployed with a launch template, then this is null.

health: Option<NodegroupHealth>

The health status of the node group. If there are issues with your node group's health, they are listed here.

update_config: Option<NodegroupUpdateConfig>

The node group update configuration.

launch_template: Option<LaunchTemplateSpecification>

If a launch template was used to create the node group, then this is the launch template that was used.

tags: Option<HashMap<String, String>>

The metadata applied to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

Implementations

impl Nodegroup

pub fn nodegroup_name(&self) -> Option<&str>

The name associated with an Amazon EKS managed node group.

pub fn nodegroup_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) associated with the managed node group.

pub fn cluster_name(&self) -> Option<&str>

The name of the cluster that the managed node group resides in.

pub fn version(&self) -> Option<&str>

The Kubernetes version of the managed node group.

pub fn release_version(&self) -> Option<&str>

If the node group was deployed using a launch template with a custom AMI, then this is the AMI ID that was specified in the launch template. For node groups that weren't deployed using a launch template, this is the version of the Amazon EKS optimized AMI that the node group was deployed with.

pub fn created_at(&self) -> Option<&Instant>

The Unix epoch timestamp in seconds for when the managed node group was created.

pub fn modified_at(&self) -> Option<&Instant>

The Unix epoch timestamp in seconds for when the managed node group was last modified.

pub fn status(&self) -> Option<&NodegroupStatus>

The current status of the managed node group.

pub fn capacity_type(&self) -> Option<&CapacityTypes>

The capacity type of your managed node group.

pub fn scaling_config(&self) -> Option<&NodegroupScalingConfig>

The scaling configuration details for the Auto Scaling group that is associated with your node group.

pub fn instance_types(&self) -> Option<&[String]>

If the node group wasn't deployed with a launch template, then this is the instance type that is associated with the node group. If the node group was deployed with a launch template, then this is null.

pub fn subnets(&self) -> Option<&[String]>

The subnets that were specified for the Auto Scaling group that is associated with your node group.

pub fn remote_access(&self) -> Option<&RemoteAccessConfig>

If the node group wasn't deployed with a launch template, then this is the remote access configuration that is associated with the node group. If the node group was deployed with a launch template, then this is null.

pub fn ami_type(&self) -> Option<&AmiTypes>

If the node group was deployed using a launch template with a custom AMI, then this is CUSTOM. For node groups that weren't deployed using a launch template, this is the AMI type that was specified in the node group configuration.

pub fn node_role(&self) -> Option<&str>

The IAM role associated with your node group. The Amazon EKS node kubelet daemon makes calls to Amazon Web Services APIs on your behalf. Nodes receive permissions for these API calls through an IAM instance profile and associated policies.

pub fn labels(&self) -> Option<&HashMap<String, String>>

The Kubernetes labels applied to the nodes in the node group.

Only labels that are applied with the Amazon EKS API are shown here. There may be other Kubernetes labels applied to the nodes in this group.

pub fn taints(&self) -> Option<&[Taint]>

The Kubernetes taints to be applied to the nodes in the node group when they are created. Effect is one of No_Schedule, Prefer_No_Schedule, or No_Execute. Kubernetes taints can be used together with tolerations to control how workloads are scheduled to your nodes.

pub fn resources(&self) -> Option<&NodegroupResources>

The resources associated with the node group, such as Auto Scaling groups and security groups for remote access.

pub fn disk_size(&self) -> Option<i32>

If the node group wasn't deployed with a launch template, then this is the disk size in the node group configuration. If the node group was deployed with a launch template, then this is null.

pub fn health(&self) -> Option<&NodegroupHealth>

The health status of the node group. If there are issues with your node group's health, they are listed here.

pub fn update_config(&self) -> Option<&NodegroupUpdateConfig>

The node group update configuration.

pub fn launch_template(&self) -> Option<&LaunchTemplateSpecification>

If a launch template was used to create the node group, then this is the launch template that was used.

pub fn tags(&self) -> Option<&HashMap<String, String>>

The metadata applied to the node group to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Node group tags do not propagate to any other resources associated with the node group, such as the Amazon EC2 instances or subnets.

impl Nodegroup

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Nodegroup