Module aws_sdk_eks::operation

Structs

AssociateEncryptionConfig

Associate encryption configuration to an existing cluster.

AssociateIdentityProviderConfig

Associate an identity provider configuration to a cluster.

CreateAddon

Creates an Amazon EKS add-on.

CreateCluster

Creates an Amazon EKS control plane.

CreateFargateProfile

Creates an AWS Fargate profile for your Amazon EKS cluster. You must have at least one Fargate

CreateNodegroup

Creates a managed node group for an Amazon EKS cluster. You can only create a node group

DeleteAddon

Delete an Amazon EKS add-on.

DeleteCluster

Deletes the Amazon EKS cluster control plane.

DeleteFargateProfile

Deletes an AWS Fargate profile.

DeleteNodegroup

Deletes an Amazon EKS node group for a cluster.

DescribeAddon

Describes an Amazon EKS add-on.

DescribeAddonVersions

Describes the Kubernetes versions that the add-on can be used with.

DescribeCluster

Returns descriptive information about an Amazon EKS cluster.

DescribeFargateProfile

Returns descriptive information about an AWS Fargate profile.

DescribeIdentityProviderConfig

Returns descriptive information about an identity provider configuration.

DescribeNodegroup

Returns descriptive information about an Amazon EKS node group.

DescribeUpdate

Returns descriptive information about an update against your Amazon EKS cluster or

DisassociateIdentityProviderConfig

Disassociates an identity provider configuration from a cluster. If you disassociate

ListAddons

Lists the available add-ons.

ListClusters

Lists the Amazon EKS clusters in your AWS account in the specified Region.

ListFargateProfiles

Lists the AWS Fargate profiles associated with the specified cluster in your AWS

ListIdentityProviderConfigs

A list of identity provider configurations.

ListNodegroups

Lists the Amazon EKS managed node groups associated with the specified cluster in your

ListTagsForResource

List the tags for an Amazon EKS resource.

ListUpdates

Lists the updates associated with an Amazon EKS cluster or managed node group in your AWS

TagResource

Associates the specified tags to a resource with the specified

UntagResource

Deletes specified tags from a resource.

UpdateAddon

Updates an Amazon EKS add-on.

UpdateClusterConfig

Updates an Amazon EKS cluster configuration. Your cluster continues to function during the

UpdateClusterVersion

Updates an Amazon EKS cluster to the specified Kubernetes version. Your cluster continues

UpdateNodegroupConfig

Updates an Amazon EKS managed node group configuration. Your node group continues to

UpdateNodegroupVersion

Updates the Kubernetes version or AMI version of an Amazon EKS managed node group.